% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSetCondenser.R
\docType{package}
\name{ConceptSetCondenser-package}
\alias{ConceptSetCondenser}
\alias{ConceptSetCondenser-package}
\title{ConceptSetCondenser: Condense concept sets to parsimonious expressions}
\description{
Condense concept sets to parsimonious expressions.
}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

}
\keyword{internal}
