% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFetch.R
\name{fetchConceptSetData}
\alias{fetchConceptSetData}
\title{Fetch data for concept set condensation}
\usage{
fetchConceptSetData(
  conceptSetExpression,
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema")
)
}
\arguments{
\item{conceptSetExpression}{The concept set expression to condense. This can
be either a JSON expression or the output of 
`ROhdsiWebApi::getConceptSetDefinition`.}

\item{connectionDetails}{The output of `DatabaseConnector::createConnectionDetails()`.
Is ignored if `connection` is provided.}

\item{connection}{A `DatabaseConnector` connection. Can be NULL if 
`connectionDetails` is provided.}

\item{cdmDatabaseSchema}{A database schema holding the OHDSI Vocabulary
tables.}

\item{tempEmulationSchema}{For database platforms that do not natively
support temp tables, a database schema where the
user has write access.}
}
\value{
A list containing the three inputs for `condenseConceptSet()`.
}
\description{
Fetch data for concept set condensation
}
