% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Condense.R
\name{condenseConceptSet}
\alias{condenseConceptSet}
\title{Condense concept set}
\usage{
condenseConceptSet(conceptSetData)
}
\arguments{
\item{conceptSetData}{An object of type `ConceptData` as created by `fetchConceptSetData`.}
}
\value{
A tibble with a representation of the optimal concept set expression
}
\description{
Condense concept set
}
